Change FirestoreCollections to both collections and subcollections, update event service
Changed FirestoreCollections to 2 different collections:

* FirestoreCollection - just the general collections

* FirestoreSubcollection - collections under a "general" collection

Created individual Firestore.ts file to hold the main adapter and have the subcollection and collection import it.

Updated EventService for the subcollection changes.
