+++ server/src/data-layer/adapters/FirestoreCollections.ts
@@ -1,5 +1,6 @@
 // credit https://plainenglish.io/blog/using-firestore-with-typescript-in-the-v9-sdk-cf36851bb099
 import "dotenv/config"
+import firestore from "./Firestore"
 import {
   Booking,
   BookingHistoryEvent,
@@ -7,28 +8,8 @@ import {
   Event,
   UserAdditionalInfo
 } from "data-layer/models/firebase"
-import { admin } from "business-layer/security/Firebase"
 
-const converter = <T>() => ({
-  toFirestore: (data: any) => data,
-  fromFirestore: (doc: any) => doc.data() as T
-})
-
-const firestore = Object.assign(
-  () => {
-    return admin.firestore()
-  },
-  {
-    doc: <T>(path: string) => {
-      return admin.firestore().doc(path).withConverter<T>(converter<T>())
-    },
-    collection: <T>(path: string) => {
-      return admin.firestore().collection(path).withConverter<T>(converter<T>())
-    }
-  }
-)
-
-const db = {
+const FirestoreCollections = {
   users: firestore.collection<UserAdditionalInfo>("users"),
   bookings: firestore.collection<Booking>("bookings"),
   bookingSlots: firestore.collection<BookingSlot>("booking_slots"),
@@ -36,4 +17,4 @@ const db = {
   events: firestore.collection<Event>("events")
 } as const
 
-export default db
+export default FirestoreCollections
